feat(rapidocr_onnxruntime): optimize code
