+++ python/rapidocr_onnxruntime/ch_ppocr_rec/utils.py
@@ -17,14 +17,14 @@ class CTCLabelDecode:
         self.dict = {char: i for i, char in enumerate(self.character)}
 
     def __call__(
-        self, preds: np.ndarray, rec_word_box: bool = False, **kwargs
+        self, preds: np.ndarray, return_word_box: bool = False, **kwargs
     ) -> List[Tuple[str, float]]:
         preds_idx = preds.argmax(axis=2)
         preds_prob = preds.max(axis=2)
         text = self.decode(
-            preds_idx, preds_prob, rec_word_box, is_remove_duplicate=True
+            preds_idx, preds_prob, return_word_box, is_remove_duplicate=True
         )
-        if rec_word_box:
+        if return_word_box:
             for rec_idx, rec in enumerate(text):
                 wh_ratio = kwargs["wh_ratio_list"][rec_idx]
                 max_wh_ratio = kwargs["max_wh_ratio"]
@@ -76,7 +76,7 @@ class CTCLabelDecode:
         self,
         text_index: np.ndarray,
         text_prob: Optional[np.ndarray] = None,
-        rec_word_box: bool = False,
+        return_word_box: bool = False,
         is_remove_duplicate: bool = False,
     ) -> List[Tuple[str, float]]:
         """convert text-index into text-label."""
@@ -91,9 +91,6 @@ class CTCLabelDecode:
             for ignored_token in ignored_tokens:
                 selection &= text_index[batch_idx] != ignored_token
 
-            char_list = [
-                self.character[text_id] for text_id in text_index[batch_idx][selection]
-            ]
             if text_prob is not None:
                 conf_list = text_prob[batch_idx][selection]
             else:
@@ -102,8 +99,11 @@ class CTCLabelDecode:
             if len(conf_list) == 0:
                 conf_list = [0]
 
+            char_list = [
+                self.character[text_id] for text_id in text_index[batch_idx][selection]
+            ]
             text = "".join(char_list)
-            if rec_word_box:
+            if return_word_box:
                 word_list, word_col_list, state_list = self.get_word_info(
                     text, selection
                 )
@@ -129,6 +129,7 @@ class CTCLabelDecode:
     ) -> Tuple[List[List[str]], List[List[int]], List[str]]:
         """
         Group the decoded characters and record the corresponding decoded positions.
+        from https://github.com/PaddlePaddle/PaddleOCR/blob/fbba2178d7093f1dffca65a5b963ec277f1a6125/ppocr/postprocess/rec_postprocess.py#L70
 
         Args:
             text: the decoded text
