+++ python/rapidocr_onnxruntime/main.py
@@ -8,6 +8,7 @@ from typing import Any, Dict, List, Optional, Tuple, Union
 import cv2
 import numpy as np
 
+from .cal_rec_boxes import CalRecBoxes
 from .ch_ppocr_cls import TextClassifier
 from .ch_ppocr_det import TextDetector
 from .ch_ppocr_rec import TextRecognizer
@@ -22,7 +23,6 @@ from .utils import (
     read_yaml,
     reduce_max_side,
     update_model_path,
-    word_box,
 )
 
 root_dir = Path(__file__).resolve().parent
@@ -61,6 +61,8 @@ class RapidOCR:
         self.max_side_len = global_config["max_side_len"]
         self.min_side_len = global_config["min_side_len"]
 
+        self.cal_rec_boxes = CalRecBoxes()
+
     def __call__(
         self,
         img_content: Union[str, np.ndarray, bytes, Path],
@@ -72,12 +74,12 @@ class RapidOCR:
         use_det = self.use_det if use_det is None else use_det
         use_cls = self.use_cls if use_cls is None else use_cls
         use_rec = self.use_rec if use_rec is None else use_rec
-        rec_word_box = False
+        return_word_box = False
         if kwargs:
             box_thresh = kwargs.get("box_thresh", 0.5)
             unclip_ratio = kwargs.get("unclip_ratio", 1.6)
             text_score = kwargs.get("text_score", 0.5)
-            rec_word_box = kwargs.get("rec_word_box", False)
+            return_word_box = kwargs.get("return_word_box", False)
             self.text_det.postprocess_op.box_thresh = box_thresh
             self.text_det.postprocess_op.unclip_ratio = unclip_ratio
             self.text_score = text_score
@@ -104,20 +106,22 @@ class RapidOCR:
             img, cls_res, cls_elapse = self.text_cls(img)
 
         if use_rec:
-            rec_res, rec_elapse = self.text_rec(img, rec_word_box)
+            rec_res, rec_elapse = self.text_rec(img, return_word_box)
 
         # fix word box by fix rotate and perspective
-        if dt_boxes is not None and rec_res is not None and rec_word_box:
-            rec_res = word_box.cal_rec_boxes(dt_boxes, img, rec_res)
-            for i, rec_res_i in enumerate(rec_res):
+        if dt_boxes is not None and rec_res is not None and return_word_box:
+            rec_res = self.cal_rec_boxes(img, dt_boxes, rec_res)
+            for rec_res_i in rec_res:
                 if rec_res_i[3]:
                     rec_res_i[3] = (
                         self._get_origin_points(rec_res_i[3], op_record, raw_h, raw_w)
                         .astype(np.int32)
                         .tolist()
                     )
+
         if dt_boxes is not None and rec_res is not None:
             dt_boxes = self._get_origin_points(dt_boxes, op_record, raw_h, raw_w)
+
         ocr_res = self.get_final_res(
             dt_boxes, cls_res, rec_res, det_elapse, cls_elapse, rec_elapse
         )
